+++ tests/algorithms/minimalistic/config_extensions/settings.testminimalistic.extra.toml
