+++ tests/algorithms/minimalistic/test_sdk_minimalistic.py
@@ -23,7 +23,7 @@ from tests.algorithms.shared import (
     TestEnum,
     NegativeZError,
 )
-from tests.algorithms.minimalistic.sample_main import main as sample_algorithm_main
+from tests.algorithms.shared import main as sample_algorithm_main
 
 os.environ["PROTEUS__AWS_REGION"] = "us-east-1"
 os.environ["PROTEUS__AWS_ENDPOINT"] = "http://localhost:9000"
