+++ tests/settings.custom.toml
@@ -6,7 +6,7 @@ algorithm_name = "hello-world"
 [runtime]
 
 # Nexus algorithm implementations to activate. At least one must be supplied
-algorithms = ["tests.algorithms.minimalistic.sample_main.TestAlgorithm"]
+algorithms = []
 
 # additional modules for the DI container
 # use these to provide custom services
